LayoutTests/editing/pasteboard/enties-api/datatransfer-items-drop-getParent-root-expected.txt

Basic test coverage for fileSystemEntry.getParent() on the root

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rootParent.name is ""
PASS rootParent.fullPath is "/"
PASS rootParent.isDirectory is true
PASS rootParent.isFile is false
PASS successfullyParsed is true

TEST COMPLETE
